requirements.txt
mcp>=1.0.0
aiohttp>=3.8.0
playwright>=1.40.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
